requirements.txt
flask
Flask-Discord
requests
discord.py
redis